move /geolocation routes into geolocationRoutes and change /users/otherGeolocations to /geolocation/active/other
